+++ src/rpc/net.cpp
@@ -142,7 +142,9 @@ static UniValue getpeerinfo(const Config &config,
                      "was an automatic/inbound connection"},
                     {RPCResult::Type::NUM, "startingheight",
                      "The starting height (block) of the peer"},
-                    {RPCResult::Type::NUM, "banscore", "The ban score"},
+                    {RPCResult::Type::NUM, "banscore",
+                     "The ban score (DEPRECATED, returned only if config "
+                     "option -deprecatedrpc=banscore is passed)"},
                     {RPCResult::Type::NUM, "synced_headers",
                      "The last header we have in common with this peer"},
                     {RPCResult::Type::NUM, "synced_blocks",
@@ -241,7 +243,10 @@ static UniValue getpeerinfo(const Config &config,
         obj.pushKV("addnode", stats.m_manual_connection);
         obj.pushKV("startingheight", stats.nStartingHeight);
         if (fStateStats) {
-            obj.pushKV("banscore", statestats.nMisbehavior);
+            if (IsDeprecatedRPCEnabled(gArgs, "banscore")) {
+                // banscore is deprecated in v0.22.11 for removal in v0.23
+                obj.pushKV("banscore", statestats.nMisbehavior);
+            }
             obj.pushKV("synced_headers", statestats.nSyncHeight);
             obj.pushKV("synced_blocks", statestats.nCommonHeight);
             UniValue heights(UniValue::VARR);